# api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"_36_Enums.ProductStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"ONLEAVE",
					"PENDING",
					"DELETED",
					"BANNED",
					"EXPIRED"
				]
			},
			"ProductStatus": {
				"$ref": "#/components/schemas/_36_Enums.ProductStatus"
			},
			"_36_Enums.OriginZone": {
				"type": "string",
				"enum": [
					"NORTH",
					"CENTRAL",
					"SOUTH",
					"MEKONG"
				]
			},
			"OriginZone": {
				"$ref": "#/components/schemas/_36_Enums.OriginZone"
			},
			"ProductListItem": {
				"properties": {
					"userId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"reputationScore": {
						"type": "number",
						"format": "double"
					},
					"originZone": {
						"allOf": [
							{
								"$ref": "#/components/schemas/OriginZone"
							}
						],
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/ProductStatus"
					},
					"price": {
						"type": "string",
						"nullable": true
					},
					"slug": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"createdAt",
					"reputationScore",
					"originZone",
					"status",
					"price",
					"slug",
					"name",
					"id"
				],
				"type": "object"
			},
			"ProductDetail": {
				"allOf": [
					{
						"$ref": "#/components/schemas/ProductListItem"
					},
					{
						"properties": {
							"tags": {
								"items": {
									"properties": {
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"name",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"images": {
								"items": {
									"properties": {
										"isPrimary": {
											"type": "boolean"
										},
										"url": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"isPrimary",
										"url",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"tagsText": {
								"type": "string",
								"nullable": true
							},
							"ver3": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"ver2": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"ver1": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"ver123": {
								"type": "string",
								"nullable": true
							},
							"oldAddr": {
								"type": "string",
								"nullable": true
							},
							"addr": {
								"type": "string",
								"nullable": true
							},
							"cId": {
								"type": "string",
								"nullable": true
							},
							"dId": {
								"type": "string",
								"nullable": true
							},
							"oldContact": {
								"type": "string",
								"nullable": true
							},
							"currentContact": {
								"type": "string",
								"nullable": true
							},
							"cliUrl": {
								"type": "string"
							},
							"cverUrl": {
								"type": "string"
							}
						},
						"required": [
							"tags",
							"images",
							"tagsText",
							"ver3",
							"ver2",
							"ver1",
							"ver123",
							"oldAddr",
							"addr",
							"cId",
							"dId",
							"oldContact",
							"currentContact",
							"cliUrl",
							"cverUrl"
						],
						"type": "object"
					}
				]
			},
			"Partial__name-string--cverUrl-string--cliUrl-string--slug-string-or-null--price-string-or-null--status-ProductStatus--originZone-OriginZone-or-null--userId-string-or-null--tagsText-string-or-null--currentContact-string-or-null--oldContact-string-or-null--dId-string-or-null--cId-string-or-null--addr-string-or-null--oldAddr-string-or-null--ver123-string-or-null--ver1-number-or-null--ver2-number-or-null--ver3-number-or-null__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"cverUrl": {
						"type": "string"
					},
					"cliUrl": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.ProductStatus"
					},
					"originZone": {
						"$ref": "#/components/schemas/_36_Enums.OriginZone"
					},
					"userId": {
						"type": "string"
					},
					"tagsText": {
						"type": "string"
					},
					"currentContact": {
						"type": "string"
					},
					"oldContact": {
						"type": "string"
					},
					"dId": {
						"type": "string"
					},
					"cId": {
						"type": "string"
					},
					"addr": {
						"type": "string"
					},
					"oldAddr": {
						"type": "string"
					},
					"ver123": {
						"type": "string"
					},
					"ver1": {
						"type": "number",
						"format": "double"
					},
					"ver2": {
						"type": "number",
						"format": "double"
					},
					"ver3": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Paste the JWT access token obtained from /auth/login"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "A description of QUIZZZ API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/products": {
			"get": {
				"operationId": "ListProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductListItem"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProductStatus"
						}
					},
					{
						"in": "query",
						"name": "originZone",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OriginZone"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tagId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductListItem"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"bearerAuth": [
							"ADMIN",
							"SELLER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ver3": {
										"type": "number",
										"format": "double",
										"nullable": true
									},
									"ver2": {
										"type": "number",
										"format": "double",
										"nullable": true
									},
									"ver1": {
										"type": "number",
										"format": "double",
										"nullable": true
									},
									"ver123": {
										"type": "string",
										"nullable": true
									},
									"oldAddr": {
										"type": "string",
										"nullable": true
									},
									"addr": {
										"type": "string",
										"nullable": true
									},
									"cId": {
										"type": "string",
										"nullable": true
									},
									"dId": {
										"type": "string",
										"nullable": true
									},
									"oldContact": {
										"type": "string",
										"nullable": true
									},
									"currentContact": {
										"type": "string",
										"nullable": true
									},
									"tagsText": {
										"type": "string",
										"nullable": true
									},
									"userId": {
										"type": "string",
										"nullable": true
									},
									"originZone": {
										"allOf": [
											{
												"$ref": "#/components/schemas/OriginZone"
											}
										],
										"nullable": true
									},
									"status": {
										"$ref": "#/components/schemas/ProductStatus"
									},
									"price": {
										"type": "string",
										"nullable": true
									},
									"slug": {
										"type": "string",
										"nullable": true
									},
									"cliUrl": {
										"type": "string"
									},
									"cverUrl": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"cliUrl",
									"cverUrl",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/products/{id}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProductDetail"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Product not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProductListItem"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Product not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"bearerAuth": [
							"ADMIN",
							"SELLER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial__name-string--cverUrl-string--cliUrl-string--slug-string-or-null--price-string-or-null--status-ProductStatus--originZone-OriginZone-or-null--userId-string-or-null--tagsText-string-or-null--currentContact-string-or-null--oldContact-string-or-null--dId-string-or-null--cId-string-or-null--addr-string-or-null--oldAddr-string-or-null--ver123-string-or-null--ver1-number-or-null--ver2-number-or-null--ver3-number-or-null__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"message"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Product not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"bearerAuth": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}